Feat/be/pet2

+++ src/main/java/company/friendsdog/dogcommunity/entity/User.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class User {
 
     // 회원가입 시 필요 정보
-    private int userNo;
+    private Long userNo;
     private String id;
     private String pwd;
     private String userName;
